Backend/requirements.txt
# ---- Core Framework ----
Flask==3.0.3
Flask-Cors==4.0.0

# ---- Model / AI / PyTorch ----
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1

# ---- Image Processing ----
Pillow==10.3.0
python-magic-bin==0.4.14

# ---- Utilities ----
numpy==1.26.4
tqdm==4.66.4

# ---- Logging & Misc ----
gunicorn==21.2.0
requests==2.31.0

# ---- Testing (optional but recommended) ----
pytest==8.2.2
